requirements.txt
diagrams>=0.23.3
PyYAML==6.0.1
